testing/mochitest/tests/Harness_sanity/test_sanity_manifest_pf.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>SimpleTest.expected = 'fail' test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
ok(true, "We expect this to pass");
ok(false, "We expect this to fail");

</script>
</body>
</html>
